fix eslint and other small stuff
