+++ applications/fishing-map/src/data/config.ts
@@ -1,3 +1,5 @@
+import { SelectOption } from '@globalfishingwatch/ui-components'
+
 export const DEFAULT_WORKSPACE = {
   latitude: 0,
   longitude: 0,
@@ -8,3 +10,26 @@ export const DEFAULT_WORKSPACE = {
 
 // Params to use replace instead of push for router history to make navigation easier
 export const REPLACE_URL_PARAMS = ['latitude', 'longitude', 'zoom']
+
+export const sources: SelectOption[] = [
+  {
+    id: 'ais',
+    label: 'AIS',
+  },
+  {
+    id: 'vms-all',
+    label: 'VMS (X COUNTRIES)',
+  },
+  {
+    id: 'vms-chile',
+    label: 'VMS Chile',
+  },
+  {
+    id: 'vms-indonesia',
+    label: 'VMS Indonesia',
+  },
+  {
+    id: 'vms-panama',
+    label: 'VMS Panama',
+  },
+]
