+++ applications/fishing-map/src/features/sidebar/Sidebar.module.css
@@ -1,4 +1,11 @@
 .sidebarHeader {
   display: flex;
   align-items: center;
+  padding-right: var(--space-S);
+}
+
+.logo {
+  width: 18rem;
+  height: auto;
+  margin-right: auto;
 }
