+++ applications/fishing-map/src/features/sidebar/VesselsSection.module.css
