+++ packages/ui-components/src/base.css
@@ -49,7 +49,7 @@
 
   /* SIZES */
   --size-XS: 2rem;
-  --size-S: 3.2rem;
+  --size-S: 2.8rem;
   --size-M: 4rem;
   --size-L: 6rem;
   --space-XS: 0.5rem;
