+++ packages/ui-components/src/tooltip/Tooltip.tsx
@@ -5,7 +5,7 @@ import styles from './Tooltip.module.css'
 function Tooltip(props: TippyProps) {
   if (!props.content) return <Fragment>{props.children}</Fragment>
   return (
-    <Tippy className={styles.tooltip} duration={100} {...props}>
+    <Tippy className={styles.tooltip} duration={100} delay={500} {...props}>
       {props.children}
     </Tippy>
   )
